feat: add modules, refactor code and add description
