+++ src/index.html
@@ -50,7 +50,6 @@
                 </header>
                 <main class="window__main main">
                     <aside class="main__sidebar sidebar">
-
                     </aside>
                     <div class='main__area'>
                         <nav class="main__tabs tabs">
