+++ src/styles/main.scss
@@ -15,10 +15,11 @@ html {
   color: white;
   margin-bottom: -5px;
   margin-top: 5px;
-  padding: 5px 17px;
+  padding: 5px 0 5px 17px;
   cursor: pointer;
   position: relative;
   z-index: 1;
+  width: 100%;
 
   &.active {
     &:after {
@@ -27,14 +28,13 @@ html {
       z-index: -1;
       top: 0;
       left: 0;
-      background: rgba(171, 171, 171, 0.42);
+      background: transparent;
       width: 100%;
       height: 25px;
       border: 1px solid orange;
       border-radius: 10px;
     }
   }
-
 }
 
 .wrapper {
@@ -155,7 +155,7 @@ code.hljs {
   height: 100%;
   min-width: 280px;
   max-width: 280px;
-  overflow: auto;
+  overflow: hidden;
   background: #808080;
   border-right: 3px solid rgb(221, 221, 221);
   padding: 5px;
@@ -232,6 +232,7 @@ code.hljs {
   width: 100%;
   height: 100%;
   position: absolute;
+  z-index: 1;
   top: 0;
   left: 0;
   background: rgba(86, 86, 86, 0.4);
@@ -317,6 +318,53 @@ code.hljs {
     background: url("../assets/images/file.png") center no-repeat;
   }
 
+}
+// description ---------- //
+.description {
+  position: absolute;
+  z-index: 1;
+  background: white;
+  border: 2px solid #000000;
+  width: 220px;
+  min-height: 50px;
+  animation: desc 1s ease;
+  padding: 10px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  &__title {
+    text-align: center;
+    margin-bottom: 5px;
+    font-size: 18px;
+    font-weight: 600;
+  }
+  &__content {
+    text-align: center;
+    margin-bottom: 5px;
+  }
+  &__edit {
+    background: transparent;
+    border: 2px solid #808080;
+    padding: 3px 10px;
+  }
+  &__area {
+    resize: none;
+    width: 100%;
+    height: 100%;
+    border: 2px solid #4d4d4d;
+    padding: 5px;
+    margin-bottom: 5px;
+  }
+}
+
+@keyframes desc {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
 }
